Add RequestLogger HTTP handler
RequestLogger has a Handle method that returns an `http.Handler`. With
configuration options to include or exclude logging the request headers,
parameters, and/or body, the handler will print to the log the request
type, the path requested, the time taken to complete the request, and
whether the request was canceled, timed out, or had another context
error.

Normal requests: Debug
Canceled and timed-out requests: Warn
Other context errors: Error